+++ purescript_waterslide/src/lib.rs
@@ -1,7 +1,6 @@
 //! This crate defines the types necessary to translate Rust data types to Purescript. Most users
-//! will only need to derive `ToPursType` with the `purescript_waterslide_derive` crate and use
+//! will only need to derive `AsPursType` with the `purescript_waterslide_derive` crate and use
 //! the `purs_module!` macro to generate modules.
-
 #![deny(missing_docs)]
 #![deny(warnings)]