+++ purescript_waterslide_derive/src/lib.rs
@@ -14,7 +14,7 @@ mod generics;
 use quote::Tokens;
 use purescript::{make_purs_constructor_impl, make_purs_type};
 
-#[proc_macro_derive(ToPursType)]
+#[proc_macro_derive(AsPursType)]
 pub fn derive_purstype(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input = input.to_string();
     let ast =
@@ -29,7 +29,7 @@ pub fn derive_purstype(input: proc_macro::TokenStream) -> proc_macro::TokenStrea
         .collect();
     let placeholder_generics_clone = placeholder_generics.clone();
 
-    let to_purs_constructor_impl = match make_purs_constructor_impl(&ast) {
+    let as_purs_constructor_impl = match make_purs_constructor_impl(&ast) {
         Ok(generated_impl) => generated_impl,
         Err(err) => panic!(
             "Could not convert the input to Purescript type constructor: {:?}",
@@ -37,25 +37,25 @@ pub fn derive_purstype(input: proc_macro::TokenStream) -> proc_macro::TokenStrea
         ),
     };
 
-    let to_purs_impl = match make_purs_type(&ast) {
+    let as_purs_impl = match make_purs_type(&ast) {
         Ok(generated_impl) => generated_impl,
         Err(err) => panic!("Could not convert the input to Purescript AST: {:?}", err),
     };
 
     let expanded = quote! {
-        impl#generics ::purescript_waterslide::ToPursConstructor for #name#generics {
-            fn to_purs_constructor() -> ::purescript_waterslide::PursConstructor {
+        impl#generics ::purescript_waterslide::AsPursConstructor for #name#generics {
+            fn as_purs_constructor() -> ::purescript_waterslide::PursConstructor {
                 #( #placeholder_generics )*
 
-                #to_purs_constructor_impl
+                #as_purs_constructor_impl
             }
         }
 
-        impl#generics ::purescript_waterslide::ToPursType for #name#generics {
-            fn to_purs_type() -> ::purescript_waterslide::PursType {
+        impl#generics ::purescript_waterslide::AsPursType for #name#generics {
+            fn as_purs_type() -> ::purescript_waterslide::PursType {
                 #( #placeholder_generics_clone )*
 
-                #to_purs_impl
+                #as_purs_impl
             }
         }
     };