+++ purescript_waterslide_tests/tests/to_purs_constructor_impls.rs
@@ -1,23 +1,23 @@
 extern crate purescript_waterslide;
 
-use purescript_waterslide::ToPursConstructor;
+use purescript_waterslide::AsPursConstructor;
 
 #[test]
 fn str_derives_as_expected() {
-    assert_eq!(&format!("{}", <&str>::to_purs_constructor()), "String");
+    assert_eq!(&format!("{}", <&str>::as_purs_constructor()), "String");
 
 }
 
 #[test]
 fn slices_derive_as_expected() {
-    assert_eq!(&format!("{}", <&[u8]>::to_purs_constructor()), "Array Int");
+    assert_eq!(&format!("{}", <&[u8]>::as_purs_constructor()), "Array Int");
 
 }
 
 #[test]
 fn vecs_derive_as_expected() {
     assert_eq!(
-        &format!("{}", <Vec<u8>>::to_purs_constructor()),
+        &format!("{}", <Vec<u8>>::as_purs_constructor()),
         "Array Int"
     );
 
@@ -25,6 +25,6 @@ fn vecs_derive_as_expected() {
 
 #[test]
 fn boxes_derive_as_expected() {
-    assert_eq!(&format!("{}", <Box<u8>>::to_purs_constructor()), "Int");
+    assert_eq!(&format!("{}", <Box<u8>>::as_purs_constructor()), "Int");
 
 }